pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "acled"
version = "0.0.1"
description = "An ACLED experiment"
authors = [
  { name = "Stefan Bojilov" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "aiohttp>=3.12.13",
    "aws-advanced-python-wrapper>=1.2.0",
    "azure-identity>=1.23.0",
    "azure-keyvault-secrets>=4.9.0",
    "boto3>=1.39.1",
    "dagster-aws>=0.27.1",
    "dagster-duckdb>=0.27.1",
    "dagster-webserver>=1.10.20",
    "ipython",
    "jupyterlab",
    "loguru",
    "matplotlib",
    "notebook",
    "numpy",
    "pandas",
    "pip",
    "polars>=1.31.0",
    "psycopg>=3.2.9",
    "psycopg2-binary>=2.9.10",
    "pytest",
    "python-dotenv",
    "ruff",
    "scikit-learn",
    "tqdm",
    "typer",
    "xlsxwriter>=3.2.5",
]
requires-python = "~=3.10.0"


[tool.ruff]
line-length = 99
src = ["acled"]
include = ["pyproject.toml", "acled/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["acled"]
force-sort-within-sections = true

